+++ saleor/graphql/attribute/mutations.py
@@ -184,10 +184,22 @@ class BaseReorderAttributeValuesMutation(BaseMutation):
         return operations
 
 
-class AttributeValueCreateInput(graphene.InputObjectType):
-    name = graphene.String(required=True, description=AttributeValueDescriptions.NAME)
+class AttributeValueInput(graphene.InputObjectType):
     value = graphene.String(description=AttributeValueDescriptions.VALUE)
     rich_text = graphene.JSONString(description=AttributeValueDescriptions.RICH_TEXT)
+    file_url = graphene.String(
+        required=False,
+        description="URL of the file attribute. Every time, a new value is created.",
+    )
+    content_type = graphene.String(required=False, description="File content type.")
+
+
+class AttributeValueCreateInput(AttributeValueInput):
+    name = graphene.String(required=True, description=AttributeValueDescriptions.NAME)
+
+
+class AttributeValueUpdateInput(AttributeValueInput):
+    name = graphene.String(required=False, description=AttributeValueDescriptions.NAME)
 
 
 class AttributeCreateInput(graphene.InputObjectType):
@@ -231,7 +243,7 @@ class AttributeUpdateInput(graphene.InputObjectType):
         description="IDs of values to be removed from this attribute.",
     )
     add_values = graphene.List(
-        AttributeValueCreateInput,
+        AttributeValueUpdateInput,
         name="addValues",
         description="New values to be created for this attribute.",
     )
@@ -257,38 +269,8 @@ class AttributeUpdateInput(graphene.InputObjectType):
 
 
 class AttributeMixin:
-    @classmethod
-    def check_values_are_unique(cls, values_input, attribute):
-        # Check values uniqueness in case of creating new attribute.
-        existing_values = attribute.values.values_list("slug", flat=True)
-        for value_data in values_input:
-            slug = slugify(value_data["name"], allow_unicode=True)
-            if slug in existing_values:
-                msg = (
-                    "Value %s already exists within this attribute."
-                    % value_data["name"]
-                )
-                raise ValidationError(
-                    {
-                        cls.ATTRIBUTE_VALUES_FIELD: ValidationError(
-                            msg, code=AttributeErrorCode.ALREADY_EXISTS
-                        )
-                    }
-                )
-
-        new_slugs = [
-            slugify(value_data["name"], allow_unicode=True)
-            for value_data in values_input
-        ]
-        if len(set(new_slugs)) != len(new_slugs):
-            raise ValidationError(
-                {
-                    cls.ATTRIBUTE_VALUES_FIELD: ValidationError(
-                        "Provided values are not unique.",
-                        code=AttributeErrorCode.UNIQUE,
-                    )
-                }
-            )
+    # must be redefined by inheriting classes
+    ATTRIBUTE_VALUES_FIELD = None
 
     @classmethod
     def clean_values(cls, cleaned_input, attribute):
@@ -320,22 +302,55 @@ class AttributeMixin:
             )
 
         is_numeric_attr = attribute_input_type == AttributeInputType.NUMERIC
+        is_swatch_attr = attribute_input_type == AttributeInputType.SWATCH
         for value_data in values_input:
-            value = value_data["name"]
-            if is_numeric_attr:
-                cls.validate_numeric_value(value)
-            slug_value = value if not is_numeric_attr else value.replace(".", "_")
-            value_data["slug"] = slugify(slug_value, allow_unicode=True)
+            cls.validate_value(attribute, value_data, is_numeric_attr, is_swatch_attr)
 
-            attribute_value = models.AttributeValue(**value_data, attribute=attribute)
-            try:
-                attribute_value.full_clean()
-            except ValidationError as validation_errors:
-                for field, err in validation_errors.error_dict.items():
-                    if field == "attribute":
-                        continue
-                    raise ValidationError({cls.ATTRIBUTE_VALUES_FIELD: err})
         cls.check_values_are_unique(values_input, attribute)
+
+    @classmethod
+    def validate_value(
+        cls,
+        attribute: models.Attribute,
+        value_data: dict,
+        is_numeric_attr: bool,
+        is_swatch_attr: bool,
+    ):
+        value = value_data["name"]
+        cls.clean_value_input_data(value_data, is_swatch_attr)
+
+        if is_numeric_attr:
+            cls.validate_numeric_value(value)
+        elif is_swatch_attr:
+            cls.validate_swatch_attr_value(value_data)
+
+        slug_value = value if not is_numeric_attr else value.replace(".", "_")
+        value_data["slug"] = slugify(slug_value, allow_unicode=True)
+
+        attribute_value = models.AttributeValue(**value_data, attribute=attribute)
+        try:
+            attribute_value.full_clean()
+        except ValidationError as validation_errors:
+            for field, err in validation_errors.error_dict.items():
+                if field == "attribute":
+                    continue
+                raise ValidationError({cls.ATTRIBUTE_VALUES_FIELD: err})
+
+    @classmethod
+    def clean_value_input_data(cls, value_data: dict, is_swatch_attr: bool):
+        swatch_fields = ["file_url", "content_type", "value"]
+        if not is_swatch_attr and any(
+            [value_data.get(field) for field in swatch_fields]
+        ):
+            raise ValidationError(
+                {
+                    cls.ATTRIBUTE_VALUES_FIELD: ValidationError(
+                        "Cannot define value, file and contentType fields "
+                        "for not swatch attribute.",
+                        code=AttributeErrorCode.INVALID.value,
+                    )
+                }
+            )
 
     @classmethod
     def validate_numeric_value(cls, value):
@@ -346,7 +361,52 @@ class AttributeMixin:
                 {
                     cls.ATTRIBUTE_VALUES_FIELD: ValidationError(
                         "Value of numeric attribute must be numeric.",
-                        code=AttributeErrorCode.INVALID,
+                        code=AttributeErrorCode.INVALID.value,
+                    )
+                }
+            )
+
+    @classmethod
+    def validate_swatch_attr_value(cls, value_data: dict):
+        if value_data.get("value") and value_data.get("file_url"):
+            raise ValidationError(
+                {
+                    cls.ATTRIBUTE_VALUES_FIELD: ValidationError(
+                        "Cannot specify both value and file for swatch attribute.",
+                        code=AttributeErrorCode.INVALID.value,
+                    )
+                }
+            )
+
+    @classmethod
+    def check_values_are_unique(cls, values_input: dict, attribute: models.Attribute):
+        # Check values uniqueness in case of creating new attribute.
+        existing_values = attribute.values.values_list("slug", flat=True)
+        for value_data in values_input:
+            slug = slugify(value_data["name"], allow_unicode=True)
+            if slug in existing_values:
+                msg = (
+                    "Value %s already exists within this attribute."
+                    % value_data["name"]
+                )
+                raise ValidationError(
+                    {
+                        cls.ATTRIBUTE_VALUES_FIELD: ValidationError(
+                            msg, code=AttributeErrorCode.ALREADY_EXISTS.value
+                        )
+                    }
+                )
+
+        new_slugs = [
+            slugify(value_data["name"], allow_unicode=True)
+            for value_data in values_input
+        ]
+        if len(set(new_slugs)) != len(new_slugs):
+            raise ValidationError(
+                {
+                    cls.ATTRIBUTE_VALUES_FIELD: ValidationError(
+                        "Provided values are not unique.",
+                        code=AttributeErrorCode.UNIQUE.value,
                     )
                 }
             )
@@ -546,7 +606,9 @@ def validate_value_is_unique(attribute: models.Attribute, value: models.Attribut
         )
 
 
-class AttributeValueCreate(ModelMutation):
+class AttributeValueCreate(AttributeMixin, ModelMutation):
+    ATTRIBUTE_VALUES_FIELD = "input"
+
     attribute = graphene.Field(Attribute, description="The updated attribute.")
 
     class Arguments:
@@ -569,7 +631,29 @@ class AttributeValueCreate(ModelMutation):
     @classmethod
     def clean_input(cls, info, instance, data):
         cleaned_input = super().clean_input(info, instance, data)
-        cleaned_input["slug"] = slugify(cleaned_input["name"], allow_unicode=True)
+        if "name" in cleaned_input:
+            cleaned_input["slug"] = slugify(cleaned_input["name"], allow_unicode=True)
+        input_type = instance.attribute.input_type
+
+        is_swatch_attr = input_type == AttributeInputType.SWATCH
+        only_swatch_fields = ["file_url", "content_type", "value"]
+        errors = {}
+        if not is_swatch_attr:
+            for field in only_swatch_fields:
+                if cleaned_input.get(field):
+                    errors[field] = ValidationError(
+                        f"The field {field} can be defined only for swatch attributes.",
+                        code=AttributeErrorCode.INVALID.value,
+                    )
+        else:
+            try:
+                cls.validate_swatch_attr_value(cleaned_input)
+            except ValidationError as error:
+                errors["value"] = error.error_dict[cls.ATTRIBUTE_VALUES_FIELD]
+                errors["fileUrl"] = error.error_dict[cls.ATTRIBUTE_VALUES_FIELD]
+        if errors:
+            raise ValidationError(errors)
+
         return cleaned_input
 
     @classmethod
@@ -590,14 +674,14 @@ class AttributeValueCreate(ModelMutation):
         return AttributeValueCreate(attribute=attribute, attributeValue=instance)
 
 
-class AttributeValueUpdate(ModelMutation):
+class AttributeValueUpdate(AttributeValueCreate):
     attribute = graphene.Field(Attribute, description="The updated attribute.")
 
     class Arguments:
         id = graphene.ID(
             required=True, description="ID of an AttributeValue to update."
         )
-        input = AttributeValueCreateInput(
+        input = AttributeValueUpdateInput(
             required=True, description="Fields required to update an AttributeValue."
         )
 
@@ -611,14 +695,16 @@ class AttributeValueUpdate(ModelMutation):
     @classmethod
     def clean_input(cls, info, instance, data):
         cleaned_input = super().clean_input(info, instance, data)
-        if "name" in cleaned_input:
-            cleaned_input["slug"] = slugify(cleaned_input["name"], allow_unicode=True)
+        if cleaned_input.get("value"):
+            cleaned_input["file_url"] = ""
+            cleaned_input["content_type"] = ""
+        elif cleaned_input.get("file_url"):
+            cleaned_input["value"] = ""
         return cleaned_input
 
     @classmethod
-    def clean_instance(cls, info, instance):
-        validate_value_is_unique(instance.attribute, instance)
-        super().clean_instance(info, instance)
+    def perform_mutation(cls, _root, info, **data):
+        return super(AttributeValueCreate, cls).perform_mutation(_root, info, **data)
 
     @classmethod
     def success_response(cls, instance):