+++ saleor/graphql/meta/tests/test_meta_mutations.py
@@ -9,6 +9,7 @@ from django.core.exceptions import ValidationError
 from ....core.error_codes import MetadataErrorCode
 from ....core.models import ModelWithMetadata
 from ....invoice.models import Invoice
+from ....payment.utils import payment_owned_by_user
 from ...tests.utils import assert_no_permission, get_graphql_content
 
 PRIVATE_KEY = "private_key"
@@ -730,6 +731,69 @@ def test_update_public_metadata_for_item_without_meta(api_client, address):
     errors = response["data"]["updateMetadata"]["errors"]
     assert errors[0]["field"] == "id"
     assert errors[0]["code"] == MetadataErrorCode.NOT_FOUND.name
+
+
+def test_update_public_metadata_for_payment_by_logged_user(
+    user_api_client, payment_with_public_metadata
+):
+    # given
+    payment_with_public_metadata.order.user = user_api_client.user
+    payment_with_public_metadata.order.save()
+    payment_id = graphene.Node.to_global_id("Payment", payment_with_public_metadata.pk)
+
+    # when
+    response = execute_update_public_metadata_for_item(
+        user_api_client, None, payment_id, "Payment", value="NewMetaValue"
+    )
+
+    # then
+    assert item_contains_proper_public_metadata(
+        response["data"]["updateMetadata"]["item"],
+        payment_with_public_metadata,
+        payment_id,
+        value="NewMetaValue",
+    )
+
+
+def test_update_public_metadata_for_payment_by_different_logged_user(
+    user2_api_client, payment_with_public_metadata
+):
+    # given
+    assert not payment_owned_by_user(
+        payment_with_public_metadata.pk, user2_api_client.user
+    )
+    payment_id = graphene.Node.to_global_id("Payment", payment_with_public_metadata.pk)
+    variables = {
+        "id": payment_id,
+        "input": [{"key": PUBLIC_KEY, "value": "NewMetaValue"}],
+    }
+
+    # when
+    response = user2_api_client.post_graphql(
+        UPDATE_PUBLIC_METADATA_MUTATION % "Payment", variables, permissions=None
+    )
+
+    # then
+    assert_no_permission(response)
+
+
+def test_update_public_metadata_for_payment_by_non_logged_user(
+    api_client, payment_with_public_metadata
+):
+    # given
+    payment_id = graphene.Node.to_global_id("Payment", payment_with_public_metadata.pk)
+    variables = {
+        "id": payment_id,
+        "input": [{"key": PUBLIC_KEY, "value": "NewMetaValue"}],
+    }
+
+    # when
+    response = api_client.post_graphql(
+        UPDATE_PUBLIC_METADATA_MUTATION % "Payment", variables, permissions=None
+    )
+
+    # then
+    assert_no_permission(response)
 
 
 def test_add_public_metadata_for_warehouse(
@@ -1715,7 +1779,7 @@ def test_staff_update_private_metadata_empty_key(
     admin_id = graphene.Node.to_global_id("User", admin_user.pk)
 
     # when
-    response = response = execute_update_private_metadata_for_item(
+    response = execute_update_private_metadata_for_item(
         staff_api_client,
         permission_manage_staff,
         admin_id,
@@ -2169,6 +2233,109 @@ def test_update_private_metadata_for_item_without_meta(api_client, address):
     errors = response["data"]["updatePrivateMetadata"]["errors"]
     assert errors[0]["field"] == "id"
     assert errors[0]["code"] == MetadataErrorCode.NOT_FOUND.name
+
+
+def test_update_private_metadata_for_payment_by_staff(
+    staff_api_client, permission_manage_payments, payment_with_private_metadata
+):
+    # given
+    payment_id = graphene.Node.to_global_id("Payment", payment_with_private_metadata.pk)
+
+    # when
+    response = execute_update_private_metadata_for_item(
+        staff_api_client,
+        permission_manage_payments,
+        payment_id,
+        "Payment",
+        value="NewMetaValue",
+    )
+
+    # then
+    assert item_contains_proper_private_metadata(
+        response["data"]["updatePrivateMetadata"]["item"],
+        payment_with_private_metadata,
+        payment_id,
+        value="NewMetaValue",
+    )
+
+
+def test_update_private_metadata_for_payment_by_app(
+    app_api_client, permission_manage_payments, payment_with_private_metadata
+):
+    # given
+    payment_id = graphene.Node.to_global_id("Payment", payment_with_private_metadata.pk)
+
+    # when
+    response = execute_update_private_metadata_for_item(
+        app_api_client,
+        permission_manage_payments,
+        payment_id,
+        "Payment",
+        value="NewMetaValue",
+    )
+
+    # then
+    assert item_contains_proper_private_metadata(
+        response["data"]["updatePrivateMetadata"]["item"],
+        payment_with_private_metadata,
+        payment_id,
+        value="NewMetaValue",
+    )
+
+
+def test_update_private_metadata_for_payment_by_staff_without_permission(
+    staff_api_client, payment_with_private_metadata
+):
+    # given
+    payment_id = graphene.Node.to_global_id("Payment", payment_with_private_metadata.pk)
+    variables = {
+        "id": payment_id,
+        "input": [{"key": PRIVATE_KEY, "value": "NewMetaValue"}],
+    }
+
+    # when
+    response = staff_api_client.post_graphql(
+        UPDATE_PRIVATE_METADATA_MUTATION % "Payment", variables, permissions=None
+    )
+
+    # then
+    assert_no_permission(response)
+
+
+def test_update_private_metadata_for_payment_by_app_without_permission(
+    app_api_client, payment
+):
+    # given
+    payment_id = graphene.Node.to_global_id("Payment", payment.pk)
+    variables = {
+        "id": payment_id,
+        "input": [{"key": PRIVATE_KEY, "value": "NewMetaValue"}],
+    }
+
+    # when
+    response = app_api_client.post_graphql(
+        UPDATE_PRIVATE_METADATA_MUTATION % "Payment", variables, permissions=None
+    )
+
+    # then
+    assert_no_permission(response)
+
+
+def test_update_private_metadata_by_customer(user_api_client, payment):
+    # given
+    payment_id = graphene.Node.to_global_id("Payment", payment.pk)
+    variables = {
+        "id": payment_id,
+        "input": [{"key": PRIVATE_KEY, "value": "NewMetaValue"}],
+    }
+
+    # when
+    response = user_api_client.post_graphql(
+        UPDATE_PRIVATE_METADATA_MUTATION % "Payment", variables, permissions=None
+    )
+
+    # then
+    assert_no_permission(response)
 
 
 def test_add_private_metadata_for_warehouse(