+++ saleor/graphql/payment/utils.py
@@ -1,0 +1,5 @@
+from typing import List
+
+
+def metadata_contains_empty_key(metadata_list: List[dict]) -> bool:
+    return not all([data["key"].strip() for data in metadata_list])