+++ CHANGELOG.md
@@ -6,9 +6,15 @@ All notable, unreleased changes to this project will be documented in this file.
 # 3.1.0 [Unreleased]
 # 3.1.0 [Unreleased]
 - Extend app by `AppExtension` - #7701 by @korycins
+- Deprecate interface field `PaymentData.reuse_source` - #7988 by @mateuszgrzyb
 - Add ExternalNotificationTrigger mutation - #7821 by @mstrumeck
 - Add Click&Collect feature - #7673 by @kuchichan
+- Introduce swatch attributes - #7261 by @IKarbowiak
 - Introduce gift card feature - #7827 by @IKarbowiak, @tomaszszymanski129
+- Deprecate `setup_future_usage` from `checkoutComplete.paymentData` input - will be removed in Saleor 4.0 - #7994 by @mateuszgrzyb
+- Possibility to pass metadata in input of `checkoutPaymentCreate` - #8076 by @mateuszgrzyb
+- Fix shipping address issue in `availableCollectionPoints` resolver for checkout - #8143 by @kuchichan
+
 
 # 3.0.0 [Unreleased]
 
@@ -149,6 +155,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Allow impersonating user by an app/staff - #7754 by @korycins:
   - Add `customerId` to `checkoutCustomerAttach` mutation
   - Add new permision `IMPERSONATE_USER`
+
 
 ### Breaking
 - Multichannel MVP: Multicurrency - #6242 by @fowczarek @d-wysocki